refactor(main.tsx): move apollo client config to module
Ref: #170
