+++ apps/mull-ui/src/apollo/apollo-client.ts
@@ -1,0 +1,34 @@
+import { ApolloClient, InMemoryCache } from '@apollo/client';
+import { setContext } from '@apollo/client/link/context';
+import { createUploadLink } from 'apollo-upload-client';
+import { getAccessToken } from '../app/access-token';
+import { environment } from '../environments/environment';
+
+const uploadLink = createUploadLink({
+  uri: `${environment.backendUrl}/graphql`,
+});
+
+const authLink = setContext((_, { headers }) => {
+  const token = getAccessToken();
+  // return the headers to the context so httpLink can read them
+  return {
+    credentials: 'include',
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+});
+
+const apolloClient = new ApolloClient({
+  link: authLink.concat(uploadLink),
+  cache: new InMemoryCache(),
+  credentials: 'include',
+  defaultOptions: {
+    watchQuery: {
+      fetchPolicy: 'network-only',
+    },
+  },
+});
+
+export default apolloClient;
