+++ apps/mull-ui/src/main.tsx
@@ -1,12 +1,9 @@
-import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
-import { setContext } from '@apollo/client/link/context';
-import { createUploadLink } from 'apollo-upload-client';
+import { ApolloProvider } from '@apollo/client';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
-import { getAccessToken } from './app/access-token';
+import apolloClient from './apollo/apollo-client';
 import App from './app/app';
-import { environment } from './environments/environment';
 
 if ('serviceWorker' in navigator) {
   window.addEventListener('load', function () {
@@ -23,36 +20,9 @@ if ('serviceWorker' in navigator) {
   });
 }
 
-const uploadLink = createUploadLink({
-  uri: `${environment.backendUrl}/graphql`,
-});
-
-const authLink = setContext((_, { headers }) => {
-  const token = getAccessToken();
-  // return the headers to the context so httpLink can read them
-  return {
-    credentials: 'include',
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : '',
-    },
-  };
-});
-
-const client = new ApolloClient({
-  link: authLink.concat(uploadLink),
-  cache: new InMemoryCache(),
-  credentials: 'include',
-  defaultOptions: {
-    watchQuery: {
-      fetchPolicy: 'network-only',
-    },
-  },
-});
-
 ReactDOM.render(
   <React.StrictMode>
-    <ApolloProvider client={client}>
+    <ApolloProvider client={apolloClient}>
       <BrowserRouter>
         <App />
       </BrowserRouter>
